Improve movie UI
