+++ ui/v2.5/src/components/Movies/MovieDetails/Movie.tsx
@@ -30,6 +30,8 @@ import {
 import { RatingStars } from "src/components/Scenes/SceneDetails/RatingStars";
 import { MovieScenesPanel } from "./MovieScenesPanel";
 import { MovieScrapeDialog } from "./MovieScrapeDialog";
+import { MovieDetailsPanel } from "./MovieDetailsPanel";
+import { MovieEditPanel } from "./MovieEditPanel";
 
 interface IMovieParams {
   id?: string;
@@ -45,6 +47,7 @@ export const Movie: React.FC = () => {
   const [isEditing, setIsEditing] = useState<boolean>(isNew);
   const [isDeleteAlertOpen, setIsDeleteAlertOpen] = useState<boolean>(false);
   const [isImageAlertOpen, setIsImageAlertOpen] = useState<boolean>(false);
+  const [movieEditState, setMovieEditState] = useState<Partial<GQL.MovieCreateInput | GQL.MovieUpdateInput>>();
 
   // Editing movie state
   const [frontImage, setFrontImage] = useState<string | undefined | null>(
@@ -53,18 +56,8 @@ export const Movie: React.FC = () => {
   const [backImage, setBackImage] = useState<string | undefined | null>(
     undefined
   );
-  const [name, setName] = useState<string | undefined>(undefined);
-  const [aliases, setAliases] = useState<string | undefined>(undefined);
-  const [duration, setDuration] = useState<number | undefined>(undefined);
-  const [date, setDate] = useState<string | undefined>(undefined);
-  const [rating, setRating] = useState<number | undefined>(undefined);
-  const [studioId, setStudioId] = useState<string>();
-  const [director, setDirector] = useState<string | undefined>(undefined);
-  const [synopsis, setSynopsis] = useState<string | undefined>(undefined);
-  const [url, setUrl] = useState<string | undefined>(undefined);
 
   // Movie state
-  const [movie, setMovie] = useState<Partial<GQL.MovieDataFragment>>({});
   const [imagePreview, setImagePreview] = useState<string | undefined>(
     undefined
   );
@@ -78,85 +71,23 @@ export const Movie: React.FC = () => {
 
   // Network state
   const { data, error, loading } = useFindMovie(id);
+  const movie = data?.findMovie;
+
   const [isLoading, setIsLoading] = useState(false);
   const [updateMovie] = useMovieUpdate();
-  const [createMovie] = useMovieCreate(getMovieInput() as GQL.MovieCreateInput);
-  const [deleteMovie] = useMovieDestroy(
-    getMovieInput() as GQL.MovieDestroyInput
-  );
-
-  const Scrapers = useListMovieScrapers();
-  const [scrapedMovie, setScrapedMovie] = useState<
-    GQL.ScrapedMovie | undefined
-  >();
-
-  const intl = useIntl();
+  const [createMovie] = useMovieCreate();
+  const [deleteMovie] = useMovieDestroy({id});
 
   // set up hotkeys
   useEffect(() => {
-    if (isEditing) {
-      Mousetrap.bind("r 0", () => setRating(NaN));
-      Mousetrap.bind("r 1", () => setRating(1));
-      Mousetrap.bind("r 2", () => setRating(2));
-      Mousetrap.bind("r 3", () => setRating(3));
-      Mousetrap.bind("r 4", () => setRating(4));
-      Mousetrap.bind("r 5", () => setRating(5));
-      // Mousetrap.bind("u", (e) => {
-      //   setStudioFocus()
-      //   e.preventDefault();
-      // });
-      Mousetrap.bind("s s", () => onSave());
-    }
-
     Mousetrap.bind("e", () => setIsEditing(true));
     Mousetrap.bind("d d", () => onDelete());
 
     return () => {
-      if (isEditing) {
-        Mousetrap.unbind("r 0");
-        Mousetrap.unbind("r 1");
-        Mousetrap.unbind("r 2");
-        Mousetrap.unbind("r 3");
-        Mousetrap.unbind("r 4");
-        Mousetrap.unbind("r 5");
-        // Mousetrap.unbind("u");
-        Mousetrap.unbind("s s");
-      }
-
       Mousetrap.unbind("e");
       Mousetrap.unbind("d d");
     };
   });
-
-  function updateMovieEditState(state: Partial<GQL.MovieDataFragment>) {
-    setName(state.name ?? undefined);
-    setAliases(state.aliases ?? undefined);
-    setDuration(state.duration ?? undefined);
-    setDate(state.date ?? undefined);
-    setRating(state.rating ?? undefined);
-    setStudioId(state?.studio?.id ?? undefined);
-    setDirector(state.director ?? undefined);
-    setSynopsis(state.synopsis ?? undefined);
-    setUrl(state.url ?? undefined);
-  }
-
-  const updateMovieData = useCallback(
-    (movieData: Partial<GQL.MovieDataFragment>) => {
-      setFrontImage(undefined);
-      setBackImage(undefined);
-      updateMovieEditState(movieData);
-      setImagePreview(movieData.front_image_path ?? undefined);
-      setBackImagePreview(movieData.back_image_path ?? undefined);
-      setMovie(movieData);
-    },
-    []
-  );
-
-  useEffect(() => {
-    if (data && data.findMovie) {
-      updateMovieData(data.findMovie);
-    }
-  }, [data, updateMovieData]);
 
   function showImageAlert(imageData: string) {
     setImageClipboard(imageData);
@@ -195,41 +126,35 @@ export const Movie: React.FC = () => {
     }
   }
 
-  function getMovieInput() {
-    const input: Partial<GQL.MovieCreateInput | GQL.MovieUpdateInput> = {
-      name,
-      aliases,
-      duration,
-      date,
-      rating: rating ?? null,
-      studio_id: studioId ?? null,
-      director,
-      synopsis,
-      url,
+  function getMovieInput(input: Partial<GQL.MovieCreateInput | GQL.MovieUpdateInput>) {
+    const ret: Partial<GQL.MovieCreateInput | GQL.MovieUpdateInput> = {
+      ...input,
       front_image: frontImage,
       back_image: backImage,
     };
 
     if (!isNew) {
-      (input as GQL.MovieUpdateInput).id = id;
+      (ret as GQL.MovieUpdateInput).id = id;
     }
-    return input;
+    return ret;
   }
 
-  async function onSave() {
+  async function onSave(input: Partial<GQL.MovieCreateInput | GQL.MovieUpdateInput>) {
     try {
       if (!isNew) {
         const result = await updateMovie({
           variables: {
-            input: getMovieInput() as GQL.MovieUpdateInput,
+            input: getMovieInput(input) as GQL.MovieUpdateInput,
           },
         });
         if (result.data?.movieUpdate) {
-          updateMovieData(result.data.movieUpdate);
           setIsEditing(false);
+          history.push(`/movies/${result.data.movieUpdate.id}`);
         }
       } else {
-        const result = await createMovie();
+        const result = await createMovie({
+          variables: getMovieInput(input) as GQL.MovieCreateInput,
+        });
         if (result.data?.movieCreate?.id) {
           history.push(`/movies/${result.data.movieCreate.id}`);
           setIsEditing(false);
@@ -261,7 +186,8 @@ export const Movie: React.FC = () => {
 
   function onToggleEdit() {
     setIsEditing(!isEditing);
-    updateMovieData(movie);
+    setFrontImage(undefined);
+    setBackImage(undefined);
   }
 
   function renderDeleteAlert() {
@@ -272,7 +198,7 @@ export const Movie: React.FC = () => {
         accept={{ text: "Delete", variant: "danger", onClick: onDelete }}
         cancel={{ onClick: () => setIsDeleteAlertOpen(false) }}
       >
-        <p>Are you sure you want to delete {name ?? "movie"}?</p>
+        <p>Are you sure you want to delete {movie?.name ?? "movie"}?</p>
       </Modal>
     );
   }
@@ -314,153 +240,19 @@ export const Movie: React.FC = () => {
     );
   }
 
-  function updateMovieEditStateFromScraper(
-    state: Partial<GQL.ScrapedMovieDataFragment>
-  ) {
-    if (state.name) {
-      setName(state.name);
-    }
-
-    if (state.aliases) {
-      setAliases(state.aliases ?? undefined);
-    }
-
-    if (state.duration) {
-      setDuration(DurationUtils.stringToSeconds(state.duration) ?? undefined);
-    }
-
-    if (state.date) {
-      setDate(state.date ?? undefined);
-    }
-
-    if (state.studio && state.studio.id) {
-      setStudioId(state.studio.id ?? undefined);
-    }
-
-    if (state.director) {
-      setDirector(state.director ?? undefined);
-    }
-    if (state.synopsis) {
-      setSynopsis(state.synopsis ?? undefined);
-    }
-    if (state.url) {
-      setUrl(state.url ?? undefined);
-    }
-
-    // image is a base64 string
-    // #404: don't overwrite image if it has been modified by the user
-    // overwrite if not new since it came from a dialog
-    // otherwise follow existing behaviour
-    if (
-      (!isNew || frontImage === undefined) &&
-      (state as GQL.ScrapedMovieDataFragment).front_image !== undefined
-    ) {
-      const imageStr = (state as GQL.ScrapedMovieDataFragment).front_image;
-      setFrontImage(imageStr ?? undefined);
-      setImagePreview(imageStr ?? undefined);
-    }
-
-    if (
-      (!isNew || backImage === undefined) &&
-      (state as GQL.ScrapedMovieDataFragment).back_image !== undefined
-    ) {
-      const imageStr = (state as GQL.ScrapedMovieDataFragment).back_image;
-      setBackImage(imageStr ?? undefined);
-      setBackImagePreview(imageStr ?? undefined);
-    }
-  }
-
-  async function onScrapeMovieURL() {
-    if (!url) return;
-    setIsLoading(true);
-
-    try {
-      const result = await queryScrapeMovieURL(url);
-      if (!result.data || !result.data.scrapeMovieURL) {
-        return;
+  function renderImage(originalImage: string | undefined | null, editedImage: string | undefined | null, altText: string) {
+    let image = originalImage;
+    if (isEditing) {
+      if (editedImage === null) {
+        image = `${image}&default=true`;
+      } else if (editedImage) {
+        image = editedImage;
       }
-
-      // if this is a new movie, just dump the data
-      if (isNew) {
-        updateMovieEditStateFromScraper(result.data.scrapeMovieURL);
-      } else {
-        setScrapedMovie(result.data.scrapeMovieURL);
-      }
-    } catch (e) {
-      Toast.error(e);
-    } finally {
-      setIsLoading(false);
-    }
-  }
-
-  function urlScrapable(scrapedUrl: string) {
-    return (
-      !!scrapedUrl &&
-      (Scrapers?.data?.listMovieScrapers ?? []).some((s) =>
-        (s?.movie?.urls ?? []).some((u) => scrapedUrl.includes(u))
-      )
-    );
-  }
-
-  function maybeRenderScrapeButton() {
-    if (!url || !isEditing || !urlScrapable(url)) {
-      return undefined;
-    }
-    return (
-      <Button
-        className="minimal scrape-url-button"
-        onClick={() => onScrapeMovieURL()}
-      >
-        <Icon icon="file-upload" />
-      </Button>
-    );
-  }
-
-  function maybeRenderScrapeDialog() {
-    if (!scrapedMovie) {
-      return;
     }
 
-    const currentMovie = getMovieInput();
-
-    // Get image paths for scrape gui
-    currentMovie.front_image = movie.front_image_path;
-    currentMovie.back_image = movie.back_image_path;
-
-    return (
-      <MovieScrapeDialog
-        movie={currentMovie}
-        scraped={scrapedMovie}
-        onClose={(m) => {
-          onScrapeDialogClosed(m);
-        }}
-      />
-    );
-  }
-
-  function onScrapeDialogClosed(p?: GQL.ScrapedMovieDataFragment) {
-    if (p) {
-      updateMovieEditStateFromScraper(p);
+    if (image) {
+      return <img alt={altText} className="logo w-50" src={image} />
     }
-    setScrapedMovie(undefined);
-  }
-
-  function onClearFrontImage() {
-    setFrontImage(null);
-    setImagePreview(
-      movie.front_image_path
-        ? `${movie.front_image_path}?default=true`
-        : undefined
-    );
-  }
-
-  function onClearBackImage() {
-    setBackImage(null);
-    setBackImagePreview(
-      movie.back_image_path
-        ? `${movie.back_image_path}?default=true`
-        : undefined
-    );
   }
 
   if (isLoading) return <LoadingIndicator />;
@@ -469,124 +261,52 @@ export const Movie: React.FC = () => {
   return (
     <div className="row">
       <div className="movie-details col">
-        {isNew && <h2>Add Movie</h2>}
         <div className="logo w-100">
           {encodingImage ? (
             <LoadingIndicator message="Encoding image..." />
           ) : (
             <>
-              <img alt={name} className="logo w-50" src={imagePreview} />
-              <img alt={name} className="logo w-50" src={backimagePreview} />
+              {renderImage(movie?.front_image_path, frontImage, "Front cover")}
+              {renderImage(movie?.back_image_path, backImage, "Back cover")}
             </>
           )}
         </div>
 
-        <Table>
-          <tbody>
-            {TableUtils.renderInputGroup({
-              title: "Name",
-              value: name ?? "",
-              isEditing: !!isEditing,
-              onChange: setName,
-            })}
-            {TableUtils.renderInputGroup({
-              title: "Aliases",
-              value: aliases,
-              isEditing,
-              onChange: setAliases,
-            })}
-            {TableUtils.renderDurationInput({
-              title: "Duration",
-              value: duration ? duration.toString() : "",
-              isEditing,
-              onChange: (value: string | undefined) =>
-                setDuration(value ? Number.parseInt(value, 10) : undefined),
-            })}
-            {TableUtils.renderInputGroup({
-              title: `Date ${isEditing ? "(YYYY-MM-DD)" : ""}`,
-              value: isEditing ? date : TextUtils.formatDate(intl, date),
-              isEditing,
-              onChange: setDate,
-            })}
-            <tr>
-              <td>Studio</td>
-              <td>
-                <StudioSelect
-                  isDisabled={!isEditing}
-                  onSelect={(items) =>
-                    setStudioId(items.length > 0 ? items[0]?.id : undefined)
-                  }
-                  ids={studioId ? [studioId] : []}
-                />
-              </td>
-            </tr>
-            {TableUtils.renderInputGroup({
-              title: "Director",
-              value: director,
-              isEditing,
-              onChange: setDirector,
-            })}
-            <tr>
-              <td>Rating</td>
-              <td>
-                <RatingStars
-                  value={rating}
-                  disabled={!isEditing}
-                  onSetRating={(value) => setRating(value)}
-                />
-              </td>
-            </tr>
-          </tbody>
-        </Table>
-
-        <Form.Group controlId="url">
-          <Form.Label>URL {maybeRenderScrapeButton()}</Form.Label>
-          <div>
-            {EditableTextUtils.renderInputGroup({
-              isEditing,
-              onChange: setUrl,
-              value: url,
-              url: TextUtils.sanitiseURL(url),
-            })}
-          </div>
-        </Form.Group>
-
-        <Form.Group controlId="synopsis">
-          <Form.Label>Synopsis</Form.Label>
-          <Form.Control
-            as="textarea"
-            readOnly={!isEditing}
-            className="movie-synopsis text-input"
-            onChange={(newValue: React.ChangeEvent<HTMLTextAreaElement>) =>
-              setSynopsis(newValue.currentTarget.value)
-            }
-            value={synopsis}
+        {!isEditing && movie ? (
+          <>
+            <MovieDetailsPanel
+              movie={movie}
+            />
+            {/* HACK - this is also rendered in the MovieEditPanel */}
+            <DetailsEditNavbar
+              objectName={movie?.name ?? "movie"}
+              isNew={isNew}
+              isEditing={isEditing}
+              onToggleEdit={onToggleEdit}
+              onSave={() => {}}
+              onImageChange={() => {}}
+              onDelete={onDelete}
+            />
+          </>
+        ) : (
+          <MovieEditPanel
+            movie={movie ?? undefined}
+            onSubmit={onSave}
+            onCancel={onToggleEdit}
+            onDelete={onDelete}
+            setFrontImage={setFrontImage}
+            setBackImage={setBackImage}
           />
-        </Form.Group>
-
-        <DetailsEditNavbar
-          objectName={name ?? "movie"}
-          isNew={isNew}
-          isEditing={isEditing}
-          onToggleEdit={onToggleEdit}
-          onSave={onSave}
-          onImageChange={onFrontImageChange}
-          onImageChangeURL={onFrontImageLoad}
-          onClearImage={onClearFrontImage}
-          onBackImageChange={onBackImageChange}
-          onBackImageChangeURL={onBackImageLoad}
-          onClearBackImage={onClearBackImage}
-          onDelete={onDelete}
-        />
+        )}
       </div>
-      {!isNew && (
+
+      {!isNew && movie && (
         <div className="col-lg-8 col-md-7">
           <MovieScenesPanel movie={movie} />
         </div>
       )}
       {renderDeleteAlert()}
       {renderImageAlert()}
-      {maybeRenderScrapeDialog()}
     </div>
   );
 };
